drm: msm: dsi-staging: hijack 60hz backlight control for 90hz
This is a hack with vibe contribution from chatgpt. We are defining
90hz's blacklight control to just redirect to 60hz's as we have noticed
does not have the issue of werid green tint when hjacking 120hz's bc.

This fixes the twitching and sudden brightness increase when switching
refresh rate to 90hz from 120hz.
It DOES NOT fix the werid colouring.

Depends on Xiaomi's drm changes.

Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>