power: qcom/smb5: Tone down temperature limits significantly
When charging with a 20W+ charger, the device sustains a very high and constant
current, and seems to not throttle the charging at all, leading to a lot of
overheating. This is certainly dangerous, and it seems like my ginkgo has
already been affected, as my modem chip got damaged and stops working randomly,
until heat is applied to the board again.

This tries to reduce the temperature limits to more conservative values,
should be ok for now.

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>